refactor: Modal -> CardCompanyModal로 네이밍 변경
재사용 불가능하다고 판단하여 네이밍 변경
